+++ Content.Goobstation.Server/Administration/Systems/GoobAdminVerbSystem.Smites.cs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 using System.Threading;
-using Content.Goobstation.Shared.MisandryBox.Smites;
+using Content.Goobstation.Shared.Smites;
 using Content.Server.Administration.Managers;
 using Content.Server.Explosion.EntitySystems;
 using Content.Shared.Administration;
