--- Content.Goobstation.Shared/MisandryBox/InputSwapComponent.cs
@@ -1,11 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-using Robust.Shared.GameObjects;
-
-namespace Content.Goobstation.Shared.MisandryBox;
-
-[RegisterComponent, NetworkedComponent]
-public sealed partial class InputSwapComponent : Component
-{
-
-}
