--- Content.Goobstation.Shared/MisandryBox/JumpScare/JumpscareMessage.cs
@@ -1,25 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-using System.IO;
-using Lidgren.Network;
-using SixLabors.ImageSharp;
-using SixLabors.ImageSharp.PixelFormats;
-
-namespace Content.Goobstation.Shared.MisandryBox.JumpScare;
-
-public sealed class JumpscareMessage : NetMessage
-{
-    public string ImagePath = "";
-
-    public override MsgGroups MsgGroup { get; } = MsgGroups.String;
-
-    public override void ReadFromBuffer(NetIncomingMessage buffer, IRobustSerializer serializer)
-    {
-        ImagePath = buffer.ReadString();
-    }
-
-    public override void WriteToBuffer(NetOutgoingMessage buffer, IRobustSerializer serializer)
-    {
-        buffer.Write(ImagePath);
-    }
-}
