+++ Content.Goobstation.Shared/Smites/InputSwapComponent.cs
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+using Robust.Shared.GameObjects;
+
+namespace Content.Goobstation.Shared.Smites;
+
+[RegisterComponent, NetworkedComponent]
+public sealed partial class InputSwapComponent : Component;
