+++ Content.Goobstation.Shared/Smites/RunWalkSwapComponent.cs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+namespace Content.Goobstation.Shared.Smites;
+
+[RegisterComponent, NetworkedComponent]
+public sealed partial class RunWalkSwapComponent : Component;
