--- Resources/Locale/en-US/_Goobstation/MisandryBox/HairSaturation.ftl
@@ -1,2 +1,0 @@
-role-timer-hair-too-neon = Your hair color is too saturated. Maximum allowed is [color=magenta]{ $threshold }[/color]. (Switch from RGB to HSV)
-role-timer-hair-not-neon = Your hair color is not saturated enough. Minimum allowed is [color=magenta] { $threshold }[/color]. (Switch from RGB to HSV)
