--- Resources/Locale/en-US/_Goobstation/MisandryBox/HairStyle.ftl
@@ -1,2 +1,0 @@
-role-timer-blacklisted-hair = Your character must not have any of the following hair styles for you to select this:
-role-timer-whitelisted-hair = Your character must have any of the following hair styles for you to select this:
