--- Resources/Locale/en-US/_Goobstation/MisandryBox/Thunderstrike.ftl
@@ -1,2 +1,0 @@
-admin-smite-thunderstrike-name = Thunderstrike
-admin-smite-thunderstrike-desc = Send EFCC on this guy
