+++ Resources/Locale/en-US/_Goobstation/administration/smites.ftl
@@ -1,3 +1,5 @@
 admin-smite-hell-teleport-name = go to hell
+admin-smite-thunderstrike-name = Thunderstrike
 
 admin-smite-hell-teleport-description = Teleports the target to hell.
+admin-smite-thunderstrike-desc = Send EFCC on this guy
