+++ Resources/Locale/en-US/_Goobstation/objectives/job.ftl
